Fixed formatting and added separate event package
