+++ ScrabblersProject/src/application/controller/GameController.java
@@ -43,6 +43,12 @@ public class GameController implements Initializable {
     @FXML
     private ListView<String> WordDisplayLV;
     
+    @FXML
+    private Button EndTurnButton;
+    
+    @FXML 
+    private Button UndoButton;
+    
     ObservableList<String> list = FXCollections.observableArrayList("This", "is", "to", "populate", "this", "listview");
 	
 	@Override
