+++ ScrabblersProject/src/application/model/GameModel.java
@@ -2,4 +2,8 @@ package application.model;
 
 public class GameModel {
 
+	public static void endTurn() {
+		
+	}  
+	
 }
